+++ src/allmydata/test/test_upload.py
@@ -28,12 +28,12 @@ from allmydata.util import log, base32
 from allmydata.util.assertutil import precondition
 from allmydata.util.deferredutil import DeferredListShouldSucceed
 from allmydata.test.no_network import GridTestMixin
-from allmydata.test.common_py3 import ShouldFailMixin
 from allmydata.storage_client import StorageFarmBroker
 from allmydata.storage.server import storage_index_to_dir
 from allmydata.client import _Client
 from .common import (
     EMPTY_CLIENT_CONFIG,
+    ShouldFailMixin,
 )
 from functools import reduce
